refactor: 패키지 구조 변경 및 Concert Service 리팩토링
1. 패키지 구조 변경
- Domain 패키지의 하위 패키지로 entity, repository를 추가했습니다.
2. Concert Service 리팩토링
- Concert Service와 Seat Service를 분리했습니다.
