+++ src/main/java/com/hhplu/hhplusconcert/app/domain/concert/repository/ConcertRepository.java
@@ -1,4 +1,4 @@
-package com.hhplu.hhplusconcert.app.domain.concert;
+package com.hhplu.hhplusconcert.app.domain.concert.repository;
 
 public interface ConcertRepository {
     void existsConcert(Long concertId);
