--- src/main/java/com/hhplu/hhplusconcert/app/domain/reservation/ReservationRepository.java
@@ -1,5 +1,0 @@
-package com.hhplu.hhplusconcert.app.domain.reservation;
-
-public interface ReservationRepository {
-    Reservation saveReservation(Reservation reservation);
-}
