+++ src/main/java/com/hhplu/hhplusconcert/app/domain/reservation/repository/ReservationRepository.java
@@ -1,0 +1,7 @@
+package com.hhplu.hhplusconcert.app.domain.reservation.repository;
+
+import com.hhplu.hhplusconcert.app.domain.reservation.entity.Reservation;
+
+public interface ReservationRepository {
+    Reservation saveReservation(Reservation reservation);
+}
