--- src/main/java/com/hhplu/hhplusconcert/app/domain/user/UserRepository.java
@@ -1,5 +1,0 @@
-package com.hhplu.hhplusconcert.app.domain.user;
-
-public interface UserRepository {
-    User getUser(Long userId);
-}
