+++ src/main/java/com/hhplu/hhplusconcert/app/domain/user/entity/User.java
@@ -1,4 +1,4 @@
-package com.hhplu.hhplusconcert.app.domain.user;
+package com.hhplu.hhplusconcert.app.domain.user.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
