+++ src/main/java/com/hhplu/hhplusconcert/app/domain/user/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package com.hhplu.hhplusconcert.app.domain.user.repository;
+
+import com.hhplu.hhplusconcert.app.domain.user.entity.User;
+
+public interface UserRepository {
+    User getUser(Long userId);
+}
