+++ src/main/java/com/hhplu/hhplusconcert/app/infrastructure/concert/ConcertJpaRepository.java
@@ -1,6 +1,6 @@
 package com.hhplu.hhplusconcert.app.infrastructure.concert;
 
-import com.hhplu.hhplusconcert.app.domain.concert.Concert;
+import com.hhplu.hhplusconcert.app.domain.concert.entity.Concert;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ConcertJpaRepository extends JpaRepository<Concert, Long> {
