+++ src/main/java/com/hhplu/hhplusconcert/app/infrastructure/concert/SeatRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.hhplu.hhplusconcert.app.infrastructure.concert;
 
-import com.hhplu.hhplusconcert.app.domain.concert.Seat;
-import com.hhplu.hhplusconcert.app.domain.concert.SeatRepository;
+import com.hhplu.hhplusconcert.app.domain.concert.entity.Seat;
+import com.hhplu.hhplusconcert.app.domain.concert.repository.SeatRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
