+++ src/main/java/com/hhplu/hhplusconcert/app/infrastructure/queue/QueueJpaRepository.java
@@ -1,6 +1,6 @@
 package com.hhplu.hhplusconcert.app.infrastructure.queue;
 
-import com.hhplu.hhplusconcert.app.domain.queue.Queue;
+import com.hhplu.hhplusconcert.app.domain.queue.entity.Queue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
