+++ src/main/java/com/hhplu/hhplusconcert/app/infrastructure/reservation/ReservationJpaRepository.java
@@ -1,6 +1,6 @@
 package com.hhplu.hhplusconcert.app.infrastructure.reservation;
 
-import com.hhplu.hhplusconcert.app.domain.reservation.Reservation;
+import com.hhplu.hhplusconcert.app.domain.reservation.entity.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReservationJpaRepository extends JpaRepository<Reservation, Long> {
