+++ src/main/java/com/hhplu/hhplusconcert/app/infrastructure/reservation/ReservationRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.hhplu.hhplusconcert.app.infrastructure.reservation;
 
-import com.hhplu.hhplusconcert.app.domain.reservation.Reservation;
-import com.hhplu.hhplusconcert.app.domain.reservation.ReservationRepository;
+import com.hhplu.hhplusconcert.app.domain.reservation.entity.Reservation;
+import com.hhplu.hhplusconcert.app.domain.reservation.repository.ReservationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
