+++ src/main/java/com/hhplu/hhplusconcert/app/infrastructure/user/UserJpaRepository.java
@@ -1,6 +1,6 @@
 package com.hhplu.hhplusconcert.app.infrastructure.user;
 
-import com.hhplu.hhplusconcert.app.domain.user.User;
+import com.hhplu.hhplusconcert.app.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
