+++ src/main/java/com/hhplu/hhplusconcert/app/infrastructure/user/UserRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.hhplu.hhplusconcert.app.infrastructure.user;
 
-import com.hhplu.hhplusconcert.app.domain.user.User;
-import com.hhplu.hhplusconcert.app.domain.user.UserRepository;
+import com.hhplu.hhplusconcert.app.domain.user.entity.User;
+import com.hhplu.hhplusconcert.app.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
