+++ src/test/java/com/hhplu/hhplusconcert/application/service/QueueServiceTest.java
@@ -1,8 +1,8 @@
 package com.hhplu.hhplusconcert.application.service;
 
 import com.hhplu.hhplusconcert.app.application.service.QueueService;
-import com.hhplu.hhplusconcert.app.domain.queue.Queue;
-import com.hhplu.hhplusconcert.app.domain.user.User;
+import com.hhplu.hhplusconcert.app.domain.queue.entity.Queue;
+import com.hhplu.hhplusconcert.app.domain.user.entity.User;
 import com.hhplu.hhplusconcert.app.infrastructure.queue.QueueJpaRepository;
 import com.hhplu.hhplusconcert.app.domain.queue.QueueStatus;
 import org.junit.jupiter.api.Nested;
